+++ src/lib/components/Details.svelte
@@ -3,7 +3,7 @@
 	lang="ts"
 >
 	import { OrientationVector } from '$lib/three/OrientationVector'
-	import { Quaternion, Vector3, MathUtils, type Vector2Like } from 'three'
+	import { Quaternion, Vector3, MathUtils, BufferAttribute } from 'three'
 
 	const vec3 = new Vector3()
 	const quaternion = new Quaternion()
@@ -13,8 +13,8 @@
 <script lang="ts">
 	import { draggable } from '@neodrag/svelte'
 	import { Check, Copy } from 'lucide-svelte'
-	import { useTask } from '@threlte/core'
-	import { Button, Icon, Select, Input } from '@viamrobotics/prime-core'
+	import { useTask, isInstanceOf } from '@threlte/core'
+	import { Button, Icon, Select, Input, Tooltip } from '@viamrobotics/prime-core'
 	import {
 		useSelectedEntity,
 		useFocusedEntity,
@@ -25,17 +25,14 @@
 	import { usePartConfig } from '$lib/hooks/usePartConfig.svelte'
 	import { FrameConfigUpdater } from '$lib/FrameConfigUpdater.svelte'
 	import { useEnvironment } from '$lib/hooks/useEnvironment.svelte'
-	import { traits, useTrait } from '$lib/ecs'
+	import { traits, useTrait, useWorld } from '$lib/ecs'
 	import { useResourceByName } from '$lib/hooks/useResourceByName.svelte'
-	import { PersistedState } from 'runed'
+	import { useCameraControls } from '$lib/hooks/useControls.svelte'
 
 	const { ...rest } = $props()
 
-	const dragPosition = new PersistedState<Vector2Like | undefined>(
-		'details-drag-position',
-		undefined
-	)
-
+	const world = useWorld()
+	const controls = useCameraControls()
 	const resourceByName = useResourceByName()
 	const frames = useFrames()
 	const partConfig = usePartConfig()
@@ -55,6 +52,7 @@
 	const box = useTrait(() => entity, traits.Box)
 	const sphere = useTrait(() => entity, traits.Sphere)
 	const capsule = useTrait(() => entity, traits.Capsule)
+	const removable = useTrait(() => entity, traits.Removable)
 
 	const framesAPI = useTrait(() => entity, traits.FramesAPI)
 	const isFrameNode = $derived(!!framesAPI.current)
@@ -233,27 +231,77 @@
 	{@const ScalarAttribute = showEditFrameOptions ? MutableField : ImmutableField}
 
 	<div
+		id="details-panel"
 		class="border-medium bg-extralight absolute top-0 right-0 z-10 m-2 {showEditFrameOptions
 			? 'w-80'
 			: 'w-60'} border p-2 text-xs"
 		use:draggable={{
 			bounds: 'body',
 			handle: dragElement,
-			defaultPosition: dragPosition.current,
-			onDragEnd(data) {
-				dragPosition.current = { x: data.offsetX, y: data.offsetY }
-			},
 		}}
 		{...rest}
 	>
 		<div class="flex items-center justify-between gap-2 pb-2">
-			<div class="flex items-center gap-1">
+			<div class="flex w-[80%] items-center gap-1">
 				<button bind:this={dragElement}>
 					<Icon name="drag" />
 				</button>
-				<strong>{name.current}</strong>
+				<strong class="overflow-hidden text-nowrap text-ellipsis">{name.current}</strong>
 				<span class="text-subtle-2">{resourceName?.subtype}</span>
 			</div>
+
+			{#if object3d}
+				<Tooltip
+					let:tooltipID
+					location="bottom"
+				>
+					<button
+						class="text-subtle-2"
+						aria-describedby={tooltipID}
+						onclick={() => {
+							const padding = 0.4
+
+							if (!controls.current) return
+
+							const { azimuthAngle, polarAngle } = controls.current
+
+							controls.current.fitToBox(object3d, true, {
+								paddingTop: padding,
+								paddingBottom: padding,
+								paddingLeft: padding,
+								paddingRight: padding,
+							})
+
+							// Preserve previous rotation
+							controls.current?.rotateAzimuthTo(azimuthAngle, true)
+							controls.current?.rotatePolarTo(polarAngle, true)
+						}}
+					>
+						<Icon name="image-filter-center-focus" />
+					</button>
+					<p slot="description">Zoom to object</p>
+				</Tooltip>
+			{/if}
+
+			{#if removable.current}
+				<Tooltip
+					let:tooltipID
+					location="bottom"
+				>
+					<button
+						class="text-subtle-2"
+						aria-describedby={tooltipID}
+						onclick={() => {
+							if (world.has(entity)) {
+								entity.destroy()
+							}
+						}}
+					>
+						<Icon name="trash-can-outline" />
+					</button>
+					<p slot="description">Remove from scene</p>
+				</Tooltip>
+			{/if}
 		</div>
 
 		<div class="border-medium -mx-2 w-[100%+0.5rem] border-b"></div>
@@ -546,6 +594,19 @@
 							},
 						})}
 					</div>
+				</div>
+			{/if}
+
+			{#if isInstanceOf(object3d, 'Points')}
+				<div>
+					<strong class="font-semibold">points</strong>
+					{@render ImmutableField({
+						label: 'count',
+						ariaLabel: 'points count',
+						value: new Intl.NumberFormat().format(
+							(object3d.geometry.getAttribute('position') as BufferAttribute).array.length / 3
+						),
+					})}
 				</div>
 			{/if}
 		</div>