+++ src/lib/hooks/useEnvironment.svelte.ts
@@ -3,7 +3,7 @@ import { getContext, setContext } from 'svelte'
 export const ENVIRONMENT_CONTEXT_KEY = Symbol('environment')
 
 interface Environemnt {
-	viewerMode: 'edit' | 'monitor'
+	viewerMode: 'edit' | 'monitor' | 'focus'
 	isStandalone: boolean
 }