+++ src/lib/hooks/useObjectEvents.svelte.ts
@@ -5,43 +5,57 @@ import { Vector2 } from 'three'
 import type { Entity } from 'koota'
 
 export const useObjectEvents = (entity: () => Entity | undefined) => {
+	const down = new Vector2()
+
 	const selectedEntity = useSelectedEntity()
 	const focusedEntity = useFocusedEntity()
 	const visibility = useVisibility()
-	const down = new Vector2()
-	const currentEntity = $derived(entity())
-
 	const cursor = useCursor()
+
+	const currentEntity = $derived(entity())
+	const visible = $derived(currentEntity ? (visibility.get(currentEntity) ?? true) : true)
+
+	const onpointerenter = (event: IntersectionEvent<MouseEvent>) => {
+		event.stopPropagation()
+		cursor.onPointerEnter()
+	}
+
+	const onpointerleave = (event: IntersectionEvent<MouseEvent>) => {
+		event.stopPropagation()
+		cursor.onPointerLeave()
+	}
+
+	const ondblclick = (event: IntersectionEvent<MouseEvent>) => {
+		event.stopPropagation()
+		focusedEntity.set(currentEntity, event.instanceId ?? event.batchId)
+	}
+
+	const onpointerdown = (event: IntersectionEvent<MouseEvent>) => {
+		down.copy(event.pointer)
+	}
+
+	const onclick = (event: IntersectionEvent<MouseEvent>) => {
+		event.stopPropagation()
+
+		if (down.distanceToSquared(event.pointer) < 0.1) {
+			selectedEntity.set(currentEntity, event.instanceId ?? event.batchId)
+		}
+	}
+
+	$effect(() => {
+		if (!visible) {
+			cursor.onPointerLeave()
+		}
+	})
 
 	return {
 		get visible() {
-			if (!currentEntity) {
-				return true
-			}
-
-			return visibility.get(currentEntity) ?? true
+			return visible
 		},
-		onpointerenter: (event: IntersectionEvent<MouseEvent>) => {
-			event.stopPropagation()
-			cursor.onPointerEnter()
-		},
-		onpointerleave: (event: IntersectionEvent<MouseEvent>) => {
-			event.stopPropagation()
-			cursor.onPointerLeave()
-		},
-		ondblclick: (event: IntersectionEvent<MouseEvent>) => {
-			event.stopPropagation()
-			focusedEntity.set(currentEntity)
-		},
-		onpointerdown: (event: IntersectionEvent<MouseEvent>) => {
-			down.copy(event.pointer)
-		},
-		onclick: (event: IntersectionEvent<MouseEvent>) => {
-			event.stopPropagation()
-
-			if (down.distanceToSquared(event.pointer) < 0.1) {
-				selectedEntity.set(currentEntity)
-			}
-		},
+		onpointerenter,
+		onpointerleave,
+		ondblclick,
+		onpointerdown,
+		onclick,
 	}
 }