+++ src/lib/loaders/pcd/worker.ts
@@ -5,7 +5,7 @@ const loader = new PCDLoader()
 export interface SuccessMessage {
 	id: number
 	positions: Float32Array<ArrayBuffer>
-	colors: Float32Array<ArrayBuffer> | null
+	colors: Uint8Array<ArrayBuffer> | null
 }
 
 export type Message =
@@ -25,8 +25,22 @@ self.onmessage = async (event) => {
 	try {
 		const pcd = loader.parse(data.buffer as ArrayBuffer)
 		if (pcd.geometry) {
-			const positions = pcd.geometry.attributes.position.array as Float32Array<ArrayBuffer>
-			const colors = (pcd.geometry.attributes.color?.array as Float32Array<ArrayBuffer>) ?? null
+			/**
+			 * Positions is _usually_ defined. However, we have experienced parsing PCDs from Viam APIs that
+			 * result in the Three.js parser not attaching this attribute, throwing errors downstream.
+			 */
+			const positions =
+				(pcd.geometry.attributes.position?.array as Float32Array<ArrayBuffer>) ??
+				new Float32Array(0)
+			const colorsFloat: Float32Array | null =
+				(pcd.geometry.attributes.color?.array as Float32Array<ArrayBuffer>) ?? null
+			const colors = colorsFloat ? new Uint8Array(colorsFloat.length) : null
+
+			if (colors) {
+				for (let i = 0, l = colorsFloat.length; i < l; i++) {
+					colors[i] = Math.round(colorsFloat[i] * 255)
+				}
+			}
 
 			postMessage(
 				{ positions, colors, id } satisfies Message,